extract `updateUnclaimedWithdrawals` to hook
